+++ constants/absences.tsx
@@ -17,18 +17,6 @@ export const TURNS =
         { id: 3, name: "Nocturno" }
     ]
 
-const ITEM_HEIGHT = 30;
-const ITEM_PADDING_TOP = 8;
-
-export const MenuProps = {
-    PaperProps: {
-        style: {
-            maxHeight: ITEM_HEIGHT * 4.5 + ITEM_PADDING_TOP,
-            width: 250,
-        },
-    },
-};
-
 export interface IAbsence {
     id: number,
     gmpId: number,
